+++ tests/testthat/setup.R
@@ -6,13 +6,17 @@ unlink(.varFile)
 
 writeLines(sqliteFile, .varFile)
 unlink(sqliteFile)
-withr::defer({
-  unlink(sqliteFile)
-  unlink(.varFile)
-}, testthat::teardown_env())
+withr::defer(
+  {
+    unlink(sqliteFile)
+    unlink(.varFile)
+  },
+  testthat::teardown_env()
+)
 
 connectionDetails <- DatabaseConnector::createConnectionDetails("sqlite",
-                                                                server = readLines(".SQLITE_PATH"))
+  server = readLines(".SQLITE_PATH")
+)
 connection <- DatabaseConnector::connect(connectionDetails)
 DatabaseConnector::disconnect(connection)
 options(rstudio.connectionObserver.errorsSuppressed = TRUE)
@@ -26,9 +30,9 @@ if (dir.exists(Sys.getenv("DATABASECONNECTOR_JAR_FOLDER"))) {
   dir.create(jdbcDriverFolder, showWarnings = FALSE)
   DatabaseConnector::downloadJdbcDrivers("postgresql", pathToDriver = jdbcDriverFolder)
   withr::defer(
-  {
-    unlink(jdbcDriverFolder, recursive = TRUE, force = TRUE)
-  },
+    {
+      unlink(jdbcDriverFolder, recursive = TRUE, force = TRUE)
+    },
     testthat::teardown_env()
   )
 }
@@ -50,22 +54,28 @@ if (Sys.getenv("CDM5_POSTGRESQL_SERVER") != "") {
     INSERT INTO test_log.test_schema_creation (schema_name) VALUES ('@testSchema');
     CREATE SCHEMA @testSchema;"
 
-  DatabaseConnector::renderTranslateExecuteSql(sql = sql,
-                                               testSchema = testSchema,
-                                               connection = testDatabaseConnection)
+  DatabaseConnector::renderTranslateExecuteSql(
+    sql = sql,
+    testSchema = testSchema,
+    connection = testDatabaseConnection
+  )
 
-  withr::defer({
-    testDatabaseConnectionCleanup <- DatabaseConnector::connect(testDatabaseConnectionDetails)
-    sql <- "
+  withr::defer(
+    {
+      testDatabaseConnectionCleanup <- DatabaseConnector::connect(testDatabaseConnectionDetails)
+      sql <- "
     DROP SCHEMA IF EXISTS @testSchema CASCADE;
     DELETE FROM test_log.test_schema_creation WHERE schema_name = '@testSchema';
     "
-    DatabaseConnector::renderTranslateExecuteSql(sql = sql,
-                                                 testSchema = testSchema,
-                                                 connection = testDatabaseConnectionCleanup)
+      DatabaseConnector::renderTranslateExecuteSql(
+        sql = sql,
+        testSchema = testSchema,
+        connection = testDatabaseConnectionCleanup
+      )
 
-    DatabaseConnector::disconnect(testDatabaseConnectionCleanup)
-    DatabaseConnector::disconnect(testDatabaseConnection)
-  }, testthat::teardown_env())
-
+      DatabaseConnector::disconnect(testDatabaseConnectionCleanup)
+      DatabaseConnector::disconnect(testDatabaseConnection)
+    },
+    testthat::teardown_env()
+  )
 }